test(svelte-query/createMutation): rename 'OnSuccessExample' to 'SuccessExample'
