+++ packages/svelte-query/tests/createMutation/SuccessExample.svelte
